+++ libs/XMP.js
@@ -3,20 +3,20 @@
 
    Description
    These functions get the font, color, or history properties used in the document from XMP.
-   See also: https://www.adobe.io/xmp/docs/
+   See also: https://developer.adobe.com/xmp/docs/
 
    Usage
    You can include this script or copy the function to use it.
 
    Notes
-   In rare cases, if you continue to use the script, it may not work.
-   In that case, restart Illustrator and try again.
+   In rare cases, the script may not work if you continue to use it.
+   In this case, restart Illustrator and try again.
 
    Requirements
    Illustrator CS or higher
 
    Version
-   2.0.0
+   2.1.0
 
    Homepage
    github.com/sky-chaser-high/adobe-illustrator-scripts
@@ -29,7 +29,7 @@
 
 /**
  * Get font properties used in the document from XMP.
- * https://www.adobe.io/xmp/docs/XMPNamespaces/xmpTPg/
+ * https://developer.adobe.com/xmp/docs/XMPNamespaces/xmpTPg/
  *
  * @param {File} src File object
  * @returns {{composite: boolean, face: string, family: string, filename: string, name: string, type: string, version: string}[]} font properties
@@ -69,7 +69,7 @@ function getFonts(src) {
 
 /**
  * Get history properties from XMP.
- * https://www.adobe.io/xmp/docs/XMPNamespaces/xmpMM/
+ * https://developer.adobe.com/xmp/docs/XMPNamespaces/xmpMM/
  *
  * @param {File} src File object
  * @returns {{action: string, parameter: string | undefined, software: string | undefined, when: Date | undefined}[]} history properties
@@ -130,7 +130,7 @@ function getHistory(src) {
 
 /**
  * Get linked file properties used in the document from XMP.
- * https://www.adobe.io/xmp/docs/XMPNamespaces/xmpMM/
+ * https://developer.adobe.com/xmp/docs/XMPNamespaces/xmpMM/
  *
  * @param {File} src File object
  * @returns {{exists: boolean, filePath: string}[]} linked file properties
@@ -170,7 +170,7 @@ function getLinkedFiles(src) {
 
 /**
  * Get plate names used in the document from XMP.
- * https://www.adobe.io/xmp/docs/XMPNamespaces/xmpTPg/
+ * https://developer.adobe.com/xmp/docs/XMPNamespaces/xmpTPg/
  *
  * @param {File} src File object
  * @returns {string[]} plate names
@@ -202,8 +202,32 @@ function getPlateNames(src) {
 
 
 /**
+ * Get ruler units used in the document from XMP.
+ * https://developer.adobe.com/xmp/docs/XMPNamespaces/XMPDataTypes/Dimensions/
+ *
+ * @param {File} src File object
+ * @returns {string} ruler units
+ */
+function getRulerUnits(src) {
+    if (ExternalObject.AdobeXMPScript == undefined) {
+        ExternalObject.AdobeXMPScript = new ExternalObject('lib:AdobeXMPScript');
+    }
+    var xmpFile = new XMPFile(src.fsName, XMPConst.FILE_UNKNOWN, XMPConst.OPEN_FOR_READ);
+    var xmpPackets = xmpFile.getXMP();
+    var xmp = new XMPMeta(xmpPackets.serialize());
+
+    var namespace = 'http://ns.adobe.com/xap/1.0/t/pg/';
+    var prop = 'xmpTPg:MaxPageSize';
+    var unit = prop + '/stDim:unit';
+
+    var ruler = xmp.getProperty(namespace, unit).value;
+    return ruler;
+}
+
+
+/**
  * Get swatch properties used in the document from XMP.
- * https://www.adobe.io/xmp/docs/XMPNamespaces/xmpTPg/
+ * https://developer.adobe.com/xmp/docs/XMPNamespaces/xmpTPg/
  *
  * @param {File} src File object
  * @returns {{colorant: